requirements.txt
tensorflow
opencv-python
ruamel.yaml
matplotlib
numpy
